// src/lib/api/generated/model/clusterServiceDeleteClusterDeletePolicy.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */

export type ClusterServiceDeleteClusterDeletePolicy = typeof ClusterServiceDeleteClusterDeletePolicy[keyof typeof ClusterServiceDeleteClusterDeletePolicy];


export const ClusterServiceDeleteClusterDeletePolicy = {
  DELETE_POLICY_UNSPECIFIED: 'DELETE_POLICY_UNSPECIFIED',
  DELETE_POLICY_DETACH_ONLY: 'DELETE_POLICY_DETACH_ONLY',
  DELETE_POLICY_CASCADE: 'DELETE_POLICY_CASCADE',
} as const;