// src/lib/api/generated/model/v1CreateMode.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */

export type V1CreateMode = typeof V1CreateMode[keyof typeof V1CreateMode];


export const V1CreateMode = {
  CREATE_MODE_UNSPECIFIED: 'CREATE_MODE_UNSPECIFIED',
  CREATE_MODE_CREATE_NEW: 'CREATE_MODE_CREATE_NEW',
  CREATE_MODE_IMPORT_EXISTING: 'CREATE_MODE_IMPORT_EXISTING',
} as const;