// src/lib/api/generated/model/v1NamespaceVisibility.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */

export type V1NamespaceVisibility = typeof V1NamespaceVisibility[keyof typeof V1NamespaceVisibility];


export const V1NamespaceVisibility = {
  NAMESPACE_VISIBILITY_UNSPECIFIED: 'NAMESPACE_VISIBILITY_UNSPECIFIED',
  NAMESPACE_VISIBILITY_PRIVATE: 'NAMESPACE_VISIBILITY_PRIVATE',
  NAMESPACE_VISIBILITY_PUBLIC: 'NAMESPACE_VISIBILITY_PUBLIC',
} as const;